resume-matcher/requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.7.0
pdfplumber>=0.11.0
python-docx>=1.1.0
ollama>=0.2.0
scikit-learn>=1.4.0
pandas>=2.2.0
plotly>=5.20.0
torch>=2.2.0
transformers>=4.40.0
